backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
python-dotenv
authlib
python-jose
requests
anyio==4.9.0
click==8.2.2
colorama==0.4.6
greenlet==3.2.3
h11==0.16.0
idna==3.10
pydantic[email]==2.11.7     
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
apscheduler
passlib
bcrypt
httpx
itsdangerous
PyJWT
bleach
gunicorn
prometheus-fastapi-instrumentator
asyncpg